requirements.txt
django 
djangorestframework 
psycopg2-binary 
python-dotenv 
django-cors-headers
dj-database-url